+++ capitalizedecorator.rb
@@ -1,0 +1,7 @@
+require_relative('./basedecorator')
+
+class CapitalizeDecorator < BaseDecorator
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
