+++ person.rb
@@ -1,4 +1,6 @@
 require_relative('./nameable')
+require_relative('./capitalizedecorator')
+require_relative('./trimmerdecorator')
 
 class Person < Nameable
   attr_accessor :name, :age
